+++ src/components/Admin/events/ListEventScreen.jsx
@@ -61,23 +61,25 @@ export const ListEventScreen = () => {
     }
 
     const handleDelete = async (id) => {
-        setloading(true);
-        await deleteEvent(id).then(response => {
-            setEvents(events.filter(event => event.uuid != id));
-            toast.success(response.data?.message);
-        }).catch(error => {
-            if(error.response?.status == 401) {
-                dispatch({
-                    type: types.logout
-                })
-                toast.error('Ha expirado la sesión');
-                navigate(app.url_login);
-            } else {
-                toast.error(error.response?.data?.message);
-            }
-            console.log(error);
-        });
-        setloading(false);
+        if (confirm("¿Estás seguro que deseas eliminar este evento?")) {
+            setloading(true);
+            await deleteEvent(id).then(response => {
+                setEvents(events.filter(event => event.uuid != id));
+                toast.success(response.data?.message);
+            }).catch(error => {
+                if(error.response?.status == 401) {
+                    dispatch({
+                        type: types.logout
+                    })
+                    toast.error('Ha expirado la sesión');
+                    navigate(app.url_login);
+                } else {
+                    toast.error(error.response?.data?.message);
+                }
+                console.log(error);
+            });
+            setloading(false);
+        }
     }
 
     useEffect(() => {
